requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
codecov==2.0.15
colorama==0.3.9
coverage==4.5.1
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.5
django-appconf==1.0.2
django-autocomplete-light==3.2.10
django-coverage==1.2.4
django-dad
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-heroku==0.3.1
django-heroku-connect
django-import-export==1.0.1
django-jet==1.0.7
django-jsonview==1.1.0
django-postgres-extra==1.20
django-smsish==1.3.2
django-tables2==1.21.2
django-bootstrap3==10.0.1
et-xmlfile==1.0.1
flaky==3.4.0
fuzzywuzzy==0.16.0
gunicorn==19.7.1
idna==2.6
jdcal==1.4
more-itertools==4.2.0
numpy==1.14.3
odfpy==1.3.6
openpyxl==2.5.3
pandas==0.19.2
pipenv==2018.5.18
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.3
pyasn1==0.4.3
pyasn1-modules==0.2.1
pytest==3.6.0
pytest-cov==2.5.1
pytest-pythonpath==0.7.2
python-dateutil==2.7.3
pytz==2018.3
PyYAML==3.12
range-key-dict==1.1.0
requests==2.19.1
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
tqdm==4.23.4
unicodecsv==0.14.1
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
